% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additive.R
\name{additive}
\alias{additive}
\title{Create a additive data}
\usage{
additive(n, samp, mu, sigma, slope)
}
\arguments{
\item{n}{number of data}

\item{samp}{number of samples}

\item{mu}{average}

\item{sigma}{sigma}

\item{slope}{slope}
}
\value{
geno and pheno type data and list
}
\description{
additive function create a additive pheno and geno type data
}
