% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjgeno.R
\name{adjgeno}
\alias{adjgeno}
\title{adjgeno}
\usage{
adjgeno(geno, evec, pca_num)
}
\arguments{
\item{geno}{geno}

\item{evec}{evec}

\item{pca_num}{pca number}
}
\value{
geno list?
}
\description{
adjgeno function is ~
}
