% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{bio_shiny}
\alias{bio_shiny}
\title{Visualization bio data through UI}
\usage{
bio_shiny()
}
\value{
UI
}
\description{
bio_shiny function allows user to adjust the visualization of your bio data through the ui
}
\examples{
    bio_shiny()
}
