% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dominant.R
\name{dominant}
\alias{dominant}
\title{Create a dominant data}
\usage{
dominant(n, samp, mu, sigma, slope)
}
\arguments{
\item{n}{number of data}

\item{samp}{number of samples}

\item{mu}{average}

\item{sigma}{sigma}

\item{slope}{slope}
}
\value{
geno and pheno type data and list
}
\description{
dominant function create a dominant pheno and geno type data
}
