% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_sample.R
\name{f_t_test}
\alias{f_t_test}
\title{Calculate t values and test}
\usage{
f_t_test(x, k)
}
\arguments{
\item{x}{x}

\item{k}{k}

\item{geno}{geno types data}

\item{pheno}{pheno types data}
}
\value{
p and p values
}
\description{
f_kendall calculate t values, kendallTest test t values
}
