% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWAS_main_sample.R
\name{GWAS_main}
\alias{GWAS_main}
\title{Use Data and Calculate}
\usage{
GWAS_main(Y, G, pc_num)
}
\arguments{
\item{Y}{phenotypes data}

\item{G}{genotypes data}

\item{pc_num}{the number of PCA}
}
\value{
many data
}
\description{
GWAS_main function use data and calculate values
}
