% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recessive.R
\name{recessive}
\alias{recessive}
\title{Create a recessive data}
\usage{
recessive(n, samp, mu, sigma, slope)
}
\arguments{
\item{n}{number of data}

\item{samp}{number of samples}

\item{mu}{average}

\item{sigma}{sigma}

\item{slope}{slope}
}
\value{
geno and pheno type data and list
}
\description{
recessive function create a recessive pheno and geno type data
}
